arreglo de las rutas del paquete, + actualizacion del gitignorepara que ya no suba imagenes de ejemplos
