--- src/main/java/Revistas/Trabajo/TrabajoApplication.java
@@ -1,13 +1,0 @@
-package Revistas.Trabajo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TrabajoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(TrabajoApplication.class, args);
-	}
-
-}
