--- src/main/java/Revistas/Trabajo/model/Rol.java
@@ -1,43 +1,0 @@
-package Revistas.Trabajo.model;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.Id;
-import jakarta.persistence.ManyToMany;
-import jakarta.persistence.Table;
-import java.util.List;
-
-@Entity
-@Table(name = "Rol", schema = "revistas")
-public class Rol {
-    @Id
-    private String id;
-    private String nombre;
-
-    @ManyToMany(mappedBy = "roles", fetch = FetchType.LAZY) 
-    private List<Usuario> usuarios; 
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getNombre() {
-        return nombre;
-    }
-
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
-
-    public List<Usuario> getUsuarios() {
-        return usuarios;
-    }
-
-    public void setUsuarios(List<Usuario> usuarios) {
-        this.usuarios = usuarios;
-    }
-}
