--- src/main/java/Revistas/Trabajo/repository/RevistaRepository.java
@@ -1,8 +1,0 @@
-package Revistas.Trabajo.repository;
-
-import Revistas.Trabajo.model.Revista;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RevistaRepository extends JpaRepository<Revista, Integer> {
-    
-}
