--- src/main/java/Revistas/Trabajo/request/AuthRequest.java
@@ -1,19 +1,0 @@
-package Revistas.Trabajo.request;
-
-public class AuthRequest {
-
-    private String username;
-    private String password;
-	public String getUsername() {
-		return username;
-	}
-	public void setUsername(String username) {
-		this.username = username;
-	}
-	public String getPassword() {
-		return password;
-	}
-	public void setPassword(String password) {
-		this.password = password;
-	}
-}
