+++ src/main/java/ar/edu/unq/gurpo2/revistas/TrabajoApplication.java
@@ -1,0 +1,13 @@
+package ar.edu.unq.gurpo2.revistas;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TrabajoApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TrabajoApplication.class, args);
+	}
+
+}
