+++ src/main/java/ar/edu/unq/gurpo2/revistas/model/Usuario_Rol.java
@@ -1,0 +1,57 @@
+package ar.edu.unq.gurpo2.revistas.model;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+
+@Entity
+@Table(name = "Usuario_Rol", schema = "revistas")
+public class Usuario_Rol {
+
+    @Id
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "id_usuario", referencedColumnName = "id") 
+    private Usuario usuario;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "id_rol", referencedColumnName = "id") 
+    private Rol rol;
+
+   
+    public Usuario_Rol() {}
+
+    public Usuario_Rol(Usuario usuario, Rol rol) {
+        this.usuario = usuario;
+        this.rol = rol;
+    }
+
+ 
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public Usuario getUsuario() {
+        return usuario;
+    }
+
+    public void setUsuario(Usuario usuario) {
+        this.usuario = usuario;
+    }
+
+    public Rol getRol() {
+        return rol;
+    }
+
+    public void setRol(Rol rol) {
+        this.rol = rol;
+    }
+}
