+++ src/main/java/ar/edu/unq/gurpo2/revistas/request/AuthRequest.java
@@ -1,0 +1,19 @@
+package ar.edu.unq.gurpo2.revistas.request;
+
+public class AuthRequest {
+
+    private String username;
+    private String password;
+	public String getUsername() {
+		return username;
+	}
+	public void setUsername(String username) {
+		this.username = username;
+	}
+	public String getPassword() {
+		return password;
+	}
+	public void setPassword(String password) {
+		this.password = password;
+	}
+}
