+++ src/main/java/ar/edu/unq/gurpo2/revistas/service/RevistaService.java
@@ -1,0 +1,26 @@
+package ar.edu.unq.gurpo2.revistas.service;
+
+import ar.edu.unq.gurpo2.revistas.model.Revista;
+import ar.edu.unq.gurpo2.revistas.repository.RevistaRepository;
+
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class RevistaService {
+
+    @Autowired
+    private RevistaRepository revistaRepository;
+
+    public Revista crearRevista(Revista nuevaRevista) {
+        return revistaRepository.save(nuevaRevista);
+    }
+
+	public List<Revista> obtenerTodasLasRevistas() {
+		return revistaRepository.findAll(); 
+    }
+		
+}
+
